+++ src/gallium/auxiliary/draw/draw_pt_fetch_shade_emit.c
@@ -251,7 +251,7 @@ static void
 fse_run(struct draw_pt_middle_end *middle,
         const unsigned *fetch_elts,
         unsigned fetch_count,
-        const ushort *draw_elts,
+        const unsigned *draw_elts,
         unsigned draw_count,
         unsigned prim_flags )
 {
@@ -309,7 +309,7 @@ static boolean
 fse_run_linear_elts(struct draw_pt_middle_end *middle,
                     unsigned start,
                     unsigned count,
-                    const ushort *draw_elts,
+                    const unsigned *draw_elts,
                     unsigned draw_count,
                     unsigned prim_flags)
 {
