+++ src/gallium/auxiliary/draw/draw_pt_fetch_shade_pipeline.c
@@ -390,7 +390,7 @@ static void
 fetch_pipeline_run(struct draw_pt_middle_end *middle,
                    const unsigned *fetch_elts,
                    unsigned fetch_count,
-                   const ushort *draw_elts,
+                   const unsigned *draw_elts,
                    unsigned draw_count,
                    unsigned prim_flags)
 {
@@ -449,7 +449,7 @@ static boolean
 fetch_pipeline_linear_run_elts(struct draw_pt_middle_end *middle,
                                unsigned start,
                                unsigned count,
-                               const ushort *draw_elts,
+                               const unsigned *draw_elts,
                                unsigned draw_count,
                                unsigned prim_flags )
 {
